Move commonly used S/R types to platform-scripture
